OTF-Tanimoto
This update adds the option to run the simulation until a target Tanimoto score is reached. The number of steps is increased dynamically during the simulation based on the current distance from the convergence criterion.

To activate this feature, add the following line to the CONTROL file:
targettani = 0.75
The value (a float) specifies the desired target Tanimoto score.

A larger initial number of steps (as set in the CONTROL file) results in a Tanimoto score that is closer to the target. While higher initial step counts may overshoot the target, this can be useful when the user is only interested in reaching a minimum Tanimoto threshold.